Better monolithic content overflow handling.
Be ruby-aware when checking if we need a break before a line. Since ruby
annotations aren't part of the line box, just checking if the
block-offset + block-size of the line box is below the fragmentation
line won't do. Include annotation overflow as well.

Also make sure that it works correctly for blocks in vertical-rl
writing-mode. This was done incorrectly in CL:2696632.

Bug: 1236056, 1178742
Change-Id: Iaa137f1e45e437d31fe93305a29d80a2153d2d1e